ARM: dts: msm: Add DT node for PIL to load zap shader
A630 uses a zap shader ucode to flush internal memories when
switching from trusted to untrusted mode. This ucode needs to be
protected and programmed by the Hypervisor. We use the PIL to
do the appropriate calls to the Hypervisor for us and it needs a
DT node to recognize what needs to be loaded.

Change-Id: Iab010b622fe2ef739b59de939385086ea3a140c9
Signed-off-by: Shrenuj Bansal <shrenujb@codeaurora.org>