ARM: dts: msm: set MinSVS as VDD_GFX min active voltage level on SDM845
Change the minimum allowed active voltage level for the VDD_GFX
rpmh-regulator device from Off to MinSVS.  This ensures that
VDD_GFX will be physically enabled whenever software requests it
to be enabled even if the minimum allowed voltage level is
requested.

Also specify MinSVS as the initial voltage level so that VDD_GFX
will be physically enabled by a software enable request even if
it occurs before a voltage level request.

Change-Id: I74457a08099b0ebf8216701a7d1e34dff9b0c1f2
Signed-off-by: David Collins <collinsd@codeaurora.org>