msm: cam: smmu: Add separate region for secondary heap for ICP
Currently, secondary heap for ICP firmware is allocated within
the shared region. This means that the properties of the secondary
heap (system cache enabled/disabled) is coupled with that of the
shared region. ICP firmware would like to be able to control the
secondary heap properties independent of the shared region. Hence,
create a separate region for secondary heap region and let ICP host
driver reserve this region during initialization. This will let the
ICP firmware to set the properties of the secondary heap independent
of the shared memory region.

Change-Id: I2e2bc4c2a79c3dfb83e6b9bd8460a47c25f240c2
Signed-off-by: Seemanta Dutta <seemanta@codeaurora.org>